apify/himanshu/storelocator/claros_com/requirements.txt
sgselenium
lxml==4.3.4
requests==2.11.1
beautifulsoup4==4.7.1
urllib3
sgrequestssglogging
